fix: resolve critical runtime bugs and compilation errors
- Create JwtRequestAttributeFilter to set institutionId/userId from JWT
  (fixes @RequestAttribute mechanism for certificate/admin/audit controllers)
- Fix SecurityConfig: add JwtRequestAttributeFilter, add institutions public URL,
  fix userDetailsService to return Spring Security UserDetails
- Fix AuthController path /v1/auth -> /api/v1/auth to match SecurityConfig
- Fix InstitutionController path /v1/institutions -> /api/v1/institutions
- Add @JsonIgnoreProperties on BaseEntity to suppress isDeleted/createdBy/updatedBy
- Fix UserRepository invalid JPQL (join through InstitutionMembership)
- Fix DashboardSnapshotRepository invalid JPQL LIMIT clause (use derived query)
- Fix grading/nursery/attendance entities: @Builder -> @SuperBuilder
- Fix ApiResponse.success() argument order in grading/attendance/nursery controllers
- Fix import mismatches (tz.elmkusoma.common.exception -> tz.elmkusoma.exception)
- Fix GradingScaleServiceImpl BigDecimal/Double type mismatches
- Fix duplicate import in GradeBoundaryServiceImpl

+++ backend/elmkusoma-core/src/main/java/tz/elmkusoma/grading/service/impl/GradingScaleServiceImpl.java
@@ -3,7 +3,7 @@ package tz.elmkusoma.grading.service.impl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import tz.elmkusoma.common.exception.ResourceNotFoundException;
+import tz.elmkusoma.exception.ResourceNotFoundException;
 import tz.elmkusoma.grading.domain.GradingScale;
 import tz.elmkusoma.grading.dto.request.CreateGradingScaleRequest;
 import tz.elmkusoma.grading.dto.response.GradeBoundaryResponse;
@@ -11,6 +11,7 @@ import tz.elmkusoma.grading.dto.response.GradingScaleResponse;
 import tz.elmkusoma.grading.repository.GradingScaleRepository;
 import tz.elmkusoma.grading.service.GradingScaleService;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -29,8 +30,8 @@ public class GradingScaleServiceImpl implements GradingScaleService {
                 .name(request.getName())
                 .description(request.getDescription())
                 .scaleType(GradingScale.ScaleType.valueOf(request.getScaleType()))
-                .minValue(request.getMinValue())
-                .maxValue(request.getMaxValue())
+                .minValue(request.getMinValue().doubleValue())
+                .maxValue(request.getMaxValue().doubleValue())
                 .isDefault(request.getIsDefault())
                 .build();
 
@@ -65,8 +66,8 @@ public class GradingScaleServiceImpl implements GradingScaleService {
         scale.setName(request.getName());
         scale.setDescription(request.getDescription());
         scale.setScaleType(GradingScale.ScaleType.valueOf(request.getScaleType()));
-        scale.setMinValue(request.getMinValue());
-        scale.setMaxValue(request.getMaxValue());
+        scale.setMinValue(request.getMinValue().doubleValue());
+        scale.setMaxValue(request.getMaxValue().doubleValue());
         scale.setIsDefault(request.getIsDefault());
 
         GradingScale saved = gradingScaleRepository.save(scale);
@@ -98,8 +99,8 @@ public class GradingScaleServiceImpl implements GradingScaleService {
                 .name(scale.getName())
                 .description(scale.getDescription())
                 .scaleType(scale.getScaleType().name())
-                .minValue(scale.getMinValue())
-                .maxValue(scale.getMaxValue())
+                .minValue(BigDecimal.valueOf(scale.getMinValue()))
+                .maxValue(BigDecimal.valueOf(scale.getMaxValue()))
                 .isDefault(scale.getIsDefault())
                 .isActive(scale.getIsActive())
                 .createdAt(scale.getCreatedAt())
